Feature - use HTTP correlation in Azure Functions
